+++ GHEtool/main_class.py
@@ -1891,36 +1891,6 @@ class Borefield(BaseClass):
         self._load.hourly_heating_load = cooling_load
         self._delete_calculated_temperatures()
 
-    def load_hourly_profile(self, file_path: str | Path, header: bool = True, separator: str = ";",
-                            decimal_seperator: str = ".", col_heating: int = 0, col_cooling: int = 1) -> None:
-        """
-        This function loads in an hourly load profile [kW].
-
-        Parameters
-        ----------
-        file_path : str
-            Path to the hourly load file
-        header : bool
-            True if this file contains a header row
-        separator : str
-            Symbol used in the file to separate the columns
-        decimal_seperator : str
-            Symbol used for the decimal number separation
-        col_heating : int
-            Column for heating
-        col_cooling : int
-            Column for cooling
-
-        Returns
-        -------
-        None
-        """
-
-        self.load = HourlyGeothermalLoad()
-
-        self.load.load_hourly_profile(file_path=file_path, header=header, separator=separator,
-                                      decimal_seperator=decimal_seperator,
-                                      col_heating=col_heating, col_cooling=col_cooling)
     @staticmethod
     def _reduce_to_monthly_load(load: list, peak: float) -> list:
         """